+++ server/main.py
@@ -29,6 +29,7 @@ from routes.system import router as system_router
 from routes.clipboard import router as clipboard_router
 from routes.files import router as files_router
 from routes.ports import router as ports_router
+from routes.push import router as push_router
 from routes.queue import router as queue_router
 
 logging.basicConfig(level=logging.INFO)
@@ -257,6 +258,7 @@ app.include_router(clipboard_router)
 app.include_router(files_router)
 app.include_router(ports_router)
 app.include_router(queue_router)
+app.include_router(push_router)
 
 
 # ---------------------------------------------------------------------------